+++ UltraCanvas/Plugins/Vector/UltraCanvasDXFReader.cpp
@@ -33,7 +33,7 @@
 // Author: UltraCanvas Framework
 
 #include "UltraCanvasCADConverters.h"
-#include "UltraCanvasVectorStorage.h"
+#include "DataFormats/UltraCanvasVectorStorage.h"
 
 #include <algorithm>
 #include <array>